create styled.ts for StoryCards component
